+++ packages/rpc_dart_transports/lib/src/transports/http2/rpc_http2_server.dart
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: MIT
 
 import 'dart:async';
-import 'dart:io';
+import 'package:universal_io/io.dart';
 
 import 'package:http2/http2.dart' as http2;
 import 'package:rpc_dart/rpc_dart.dart';