+++ packages/rpc_dart_transports/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.1.2
+- fix: websocket connection in flutter web
+
+## 1.1.1
+- fix: web io import - remove websocket transport
+
+## 1.1.0
+
+- web: add universal io support to use on web
+
 ## 1.0.1
 
 - deps: update rpc_dart